+++ osu.Game/osu.Game.csproj
@@ -94,10 +94,8 @@
     <Compile Include="Graphics\UserInterface\OsuContextMenuItem.cs" />
     <Compile Include="Input\Binding.cs" />
     <Compile Include="Input\BindingStore.cs" />
-    <Compile Include="Input\GlobalHotkeys.cs" />
     <Compile Include="Input\IHandleActions.cs" />
     <Compile Include="Input\KeyCombination.cs" />
-    <Compile Include="Input\GlobalAction.cs" />
     <Compile Include="Input\GlobalActionMappingInputManager.cs" />
     <Compile Include="IO\FileStore.cs" />
     <Compile Include="IO\FileInfo.cs" />
